Opcode containers vc4 now fully 64-bits; all unit tests passing vc4

+++ Lib/vc4/Invoke.h
@@ -6,7 +6,7 @@
 
 namespace V3DLib {
 
-using Code   = SharedArray<uint32_t>;
+using Code   = SharedArray<uint64_t>;
 using Data   = SharedArray<uint32_t>;
 
 void invoke(int numQPUs, Code &codeMem, IntList const &params, Data &uniforms, Data &launch_messages);
